move token to new package
